Modify AuthenticationService.register to return String
Returning `RegistrationResponseBody` from the authentication service
makes little sense; a service should not have anything to do with any of
the components of http request or response. This new api is better
because it returns a string simply describing what went wrong. The
controller can then analyze the string to figure out what exactly went
wrong and respond appropriately.